Restore processing for NA formats
An earlier commit (040628bc56a5, 'Drop processing for
"NA" metadata items') removed 'NA' processing but went
too far.  PhoneNumberDesc objects with NA in them are
indeed gone, but an NA can still occur in an <intlFormat>
element of the XML.  This should map to a missing intl_format
for the relevant pattern.

Fixes #96.